+++ core/argv.js
@@ -126,10 +126,10 @@ Argv.prototype.execDefinedOptions = function () {
 Argv.prototype._showHelp = function (error) {
 	var package = require('../package.json');
 	console.log('\n' + package.name + ':', package.description);
-	console.log('Authored by', package.author);
-	console.log('Version:', package.version);
-	console.log('Repository:', package.repository.url);
-	console.log('\ngracenode command help:\n');
+	console.log('\nAuthored by', package.author);
+	console.log('\nVersion:', package.version);
+	console.log('\nRepository:', package.repository.url);
+	console.log('\nOptions:');
 	for (var i = 0, len = defKeys.length; i < len; i++) {
 		var arg = defKeys[i];
 		var spaces = this._createSpaces(this._maxLen - arg.length);