+++ scripts/lint/lint.sh
@@ -186,7 +186,7 @@ log "" "root path: $path";
 log "yellow" "$MARK lint the given directories/files";
 
 if [ ${#dirList} -eq 0 ]; then
-	log "yellow" "$MARK no directories/files to lint";
+	log "yellow" "$MARK no directories/files to lint: use -a option to give directories/files to lint. see --help for more detail";
 fi
 
 for item in "${dirList[@]}"; do