+++ scripts/daemon/restart.js
@@ -1,4 +1,3 @@
-var exec = require('child_process').exec;
 var gn = require('gracenode');
 var logger = gn.log.create('daemon-restart');
 var lib = require('./utils/lib');
@@ -6,7 +5,7 @@ var talk = require('./utils/talk');
 
 module.exports = function (path) {
 	// listener for exceptions
-	gn.on('uncaughtException', function (error) {
+	gn.on('uncaughtException', function () {
 		logger.error(lib.color(path, lib.COLORS.RED));
 		gn.exit();
 	});