+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+This is a list of manually mantained changes nad updates for each version.
+
 ## Version 1.2.2
 
 ## Added